# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bridge-LFU API",
    "version": "1.0.0",
    "description": "Bridge LFU (Licence Follow-Up) — API de suivi des licences logicielles et de l'obsolescence des équipements pour les clients de Bridge.\n\nFonctionnalités principales:\n- Authentification et gestion des utilisateurs (rôles: admin, technicien, client)\n- Gestion des clients, licences, équipements, fournisseurs, types\n- Notifications et alertes (échéances licences, obsolescence équipements)\n- Tableaux de bord et statistiques\n- Rapports exportables (JSON/CSV/PDF/Excel)\n\nConventions d’API:\n- Authentification: session via Supabase (cookies). Lancez d'abord POST /api/auth/login dans le même navigateur/onglet.\n- Pagination: paramètres `page` et `limit`, réponses standard `{ data, count, page, totalPages, hasMore }`.\n- Filtres: exposés par ressource (ex: `status`, `client_id`, `type_id`, `date_from`, `date_to`).\n- Exports: utilisez le paramètre `format` (json|csv|pdf|excel) sur les endpoints de rapports.\n\nSupport: contactez l'équipe Bridge pour l’onboarding et les accès."
  },
  "servers": [
    { "url": "/", "description": "Serveur courant" }
  ],
...
  "paths": {
...
    "/api/licenses": {
      "get": {
        "tags": ["Licenses"],
        "summary": "Lister les licences",
        "parameters": [
          { "in": "query", "name": "page", "schema": { "type": "integer", "default": 1 }},
          { "in": "query", "name": "limit", "schema": { "type": "integer", "default": 10 }},
          { "in": "query", "name": "search", "schema": { "type": "string" }},
          { "in": "query", "name": "client_id", "schema": { "type": "string", "format": "uuid" }},
          { "in": "query", "name": "status", "schema": { "$ref": "#/components/schemas/LicenseStatus" }},
          { "in": "query", "name": "editor", "schema": { "type": "string" }},
          { "in": "query", "name": "type_id", "schema": { "type": "string", "format": "uuid" }},
          { "in": "query", "name": "expiry_date_start", "schema": { "type": "string", "format": "date" }},
          { "in": "query", "name": "expiry_date_end", "schema": { "type": "string", "format": "date" }}
        ],
        "responses": {
          "200": {
            "description": "Liste paginée des licences",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaginatedLicenseResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Licenses"],
        "summary": "Créer une licence",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LicenseCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Licence créée",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/License" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
    "/api/licenses/{licenseId}/status": {
      "get": { "tags": ["Licenses"], "summary": "Statut d'une licence", "parameters": [ { "in": "path", "name": "licenseId", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "responses": { "200": { "description": "Statut" } } }
    },
...
    "/api/licenses/export": { "get": { "tags": ["Licenses"], "summary": "Exporter les licences", "parameters": [ { "in": "query", "name": "format", "schema": { "type": "string", "enum": ["csv", "xlsx", "pdf"], "default": "csv" } } ], "responses": { "200": { "description": "Fichier export" } } } },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "message": { "type": "string" }
        }
      },
...
      "License": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "editor": { "type": "string", "nullable": true },
          "version": { "type": "string", "nullable": true },
          "license_key": { "type": "string", "nullable": true },
          "purchase_date": { "type": "string", "format": "date", "nullable": true },
          "expiry_date": { "type": "string", "format": "date" },
          "cost": { "type": "number", "nullable": true },
          "client_id": { "type": "string", "format": "uuid" },
          "type_id": { "type": "string", "format": "uuid", "nullable": true },
          "supplier_id": { "type": "string", "format": "uuid", "nullable": true },
          "description": { "type": "string", "nullable": true },
          "status": { "$ref": "#/components/schemas/LicenseStatus" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "LicenseView": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "editor": { "type": "string", "nullable": true },
          "version": { "type": "string", "nullable": true },
          "license_key": { "type": "string", "nullable": true },
          "purchase_date": { "type": "string", "format": "date", "nullable": true },
          "expiry_date": { "type": "string", "format": "date" },
          "cost": { "type": "number", "nullable": true },
          "client_id": { "type": "string", "format": "uuid" },
          "type_id": { "type": "string", "format": "uuid", "nullable": true },
          "supplier_id": { "type": "string", "format": "uuid", "nullable": true },
          "description": { "type": "string", "nullable": true },
          "status": { "$ref": "#/components/schemas/LicenseStatus" },
          "created_at": { "type": "string", "format": "date-time" },
          "client_name": { "type": "string", "nullable": true },
          "client_email": { "type": "string", "format": "email", "nullable": true },
          "created_by_name": { "type": "string", "nullable": true },
          "supplier_name": { "type": "string", "nullable": true },
          "type_name": { "type": "string", "nullable": true }
        }
      },
      "LicenseCreate": {
        "type": "object",
        "required": ["name", "expiryDate", "clientId"],
        "properties": {
          "name": { "type": "string" },
          "editor": { "type": "string" },
          "version": { "type": "string" },
          "licenseKey": { "type": "string" },
          "purchaseDate": { "type": "string", "format": "date" },
          "expiryDate": { "type": "string", "format": "date" },
          "cost": { "type": "number" },
          "clientId": { "type": "string", "format": "uuid" },
          "typeId": { "type": "string", "format": "uuid" },
          "supplierId": { "type": "string", "format": "uuid" },
          "description": { "type": "string" }
        }
      },
      "LicenseStatus": {
        "type": "string",
        "enum": ["active", "expired", "about_to_expire", "cancelled"]
      }
...
      "PaginatedLicenseResponse": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/LicenseView" } },
          "count": { "type": "integer" },
          "page": { "type": "integer" },
          "totalPages": { "type": "integer" },
          "hasMore": { "type": "boolean" }
        }
      },
...
    },
    "responses": {
...
      "BadRequest": {
        "description": "Requête invalide",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Unauthorized": {
        "description": "Non authentifié",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Forbidden": {
        "description": "Permissions insuffisantes",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    }
  }
}